Included mechanic where players are not able to use items on hands when running
