+++ Game_LLM/Assets/Scripts/Player/hands.cs.meta
